reviews/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async==0.6.2
async-timeout==3.0.1
attrs==21.2.0
Baker==1.3
beautifulsoup4==4.9.3
bs4==0.0.1
celery==3.1.18
colorama==0.4.4
Django==3.2.6
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-SQLAlchemy==2.5.1
jsonschema==3.2.0
model-bakery==1.3.2
numpy==1.21.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0
PyYAML==5.4.1
requests==2.26.0
salary==2.0
selenium==3.141.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==20.4.6
